+++ v1projectBatchFinal(1)done/v1projectBatchFinal(1)/v1project/src/main/java/com/example/v1project/service/BatchService.java
@@ -8,5 +8,5 @@ public interface BatchService {
 
     List<Batches> getAllBatches();
     Batches createBatch(Batches batch);
-    void deleteBatchById(int batchId);
+    void deleteBatchById(long batchId);
 }
