+++ v1projectBatchFinal(1)done/v1projectBatchFinal(1)/v1project/src/main/java/com/example/v1project/service/BatchServiceImpl.java
@@ -20,7 +20,7 @@ public class BatchServiceImpl {
         return batchDao.findAll();
     }
 
-    public Batches getBatchById(int batchId) {
+    public Batches getBatchById(long batchId) {
         Optional<Batches> optionalBatch = batchDao.findById(batchId);
         return optionalBatch.orElse(null);
     }
@@ -35,7 +35,7 @@ public class BatchServiceImpl {
         return batchDao.save(batch);
     }
 
-    public void deleteBatchById(int batchId) {
+    public void deleteBatchById(long batchId) {
         batchDao.deleteById(batchId);
     }
